Moved aof and snapshot packages from server folder to src folder. Created new package for evictions
