+++ src/server/server.go
@@ -6,10 +6,10 @@ import (
 	"crypto/x509"
 	"errors"
 	"fmt"
+	"github.com/echovault/echovault/src/aof"
 	"github.com/echovault/echovault/src/memberlist"
 	"github.com/echovault/echovault/src/raft"
-	"github.com/echovault/echovault/src/server/aof"
-	"github.com/echovault/echovault/src/server/snapshot"
+	"github.com/echovault/echovault/src/snapshot"
 	"github.com/echovault/echovault/src/utils"
 	"io"
 	"log"
